% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{css}
\alias{css}
\title{Testing of cubic smoothing splines and semi-parametric regression models}
\usage{
css(y, x, U = NULL)
}
\arguments{
\item{y}{an \eqn{n}-dimensional vector of responses.}

\item{x}{an \eqn{n}-dimensional vector of measuring points.}

\item{U}{a full rank \eqn{n\times k} matrix of values of \eqn{k} explanatory variables.
If the matrix \eqn{U} is given, the tests for cubic smoothing spline (test 1), semi-parametric model (test 2) and linear model (test 3)
are performed.  If the matrix \eqn{U} is not given, only the test for cubic
smoothing spline (test 1) is performed.}
}
\value{
A list containing the following components:
\describe{
\item{y}{\eqn{n}-dimensional vector of responses.}
\item{x}{\eqn{n}-dimensional vector of measuring points.}
\item{alpha}{the value of \eqn{\alpha}.}
\item{c}{estimated number of eigenvectors.}
\item{gcv}{values of the gcv criteria.}
\item{ghat}{estimate of cubic smoothing spline.}
\item{F.value1}{observed value of the F test statistic (Test 1).}
\item{df1.1}{degrees of freedom of the numerator of the F test statistic (Test 1).}
\item{df2.1}{degrees of freedom of the denominator of the F test statistic (Test 1).}
\item{p.value1}{p-value (Test 1).}
\item{F.value2}{observed value of the F test statistic (Test 2).}
\item{df1.2}{degrees of freedom of the numerator of the F test statistic (Test 2).}
\item{df2.2}{degrees of freedom of the denominator of the F test statistic (Test 2).}
\item{p.value2}{p-value (Test 2).}
\item{F.value3}{observed value of the F test statistic (Test 3).}
\item{df1.3}{degrees of freedom of the numerator of the F test statistic (Test 3).}
\item{df2.3}{degrees of freedom of the denominator of the F test statistic (Test 3).}
\item{p.value3}{p-value (Test 3).}
}
}
\description{
\code{css} is used to test the cubic smoothing splines and semi-parametric regression models.
}
\details{
Here are the details of the function...
}
