% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gcv.alpha}
\alias{gcv.alpha}
\title{Generalized cross-validation}
\usage{
gcv.alpha(Y, x, A, alpha.min, alpha.max, len = 20)
}
\arguments{
\item{Y}{a \eqn{q \times n} matrix of independent \eqn{q \times 1} response vectors.}

\item{x}{a \code{q}-dimensional vector of the time points.}

\item{A}{an \eqn{n \times m} between-individual design matrix.}

\item{alpha.min}{the minimum value of the grid of \eqn{\alpha}s.}

\item{alpha.max}{the maximum value of the grid of \eqn{\alpha}s.}

\item{len}{the number of values in the grid of \eqn{\alpha}s.}
}
\value{
A list containing the following components:
\describe{
\item{gcv}{a vector of gcv values.}
\item{alpha}{a vector of alpha values.}
\item{alpha.hat}{the alpha value that corresponds to the minimum of gcv.}
\item{ind.min}{the index of alpha that corresponds to the minimum of gcv.}
\item{ind.min}{the index of alpha that corresponds to the minimum of gcv.}
\item{minpoint}{a logical value. If TRUE, the minimum of gcv was found at
\eqn{alpha[ind.min]}, where \eqn{1<ind.min<length(alpha)}. If FALSE, the
minimum of gcv was found at \eqn{alpha[1]} or \eqn{alpha[length(alpha)]}.}
}
}
\description{
\code{gcv.alpha} estimates the smoothing parameter \eqn{\alpha} by using generalized cross-validation criteria
}
\details{
Here are the details of the function...
}
\keyword{internal}
