% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gcv.dim}
\alias{gcv.dim}
\title{Generalized cross-validation}
\usage{
gcv.dim(Y, A, M)
}
\arguments{
\item{Y}{a \eqn{q \times n} matrix of independent \eqn{q \times 1} response vectors.}

\item{A}{an \eqn{n \times m} between-individual design matrix.}

\item{M}{matrix of eigenvectors of \code{S}.}
}
\value{
the number of eigenvectors.
}
\description{
\code{gcv.dim} estimates the number of eigenvectors by using generalized cross-validation criteria
}
\details{
Here are the details of the function...
}
\keyword{internal}
