% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gcv1.dim}
\alias{gcv1.dim}
\title{Generalized cross-validation (cubic smoothing splines)}
\usage{
gcv1.dim(y, Tn)
}
\arguments{
\item{Y}{a \eqn{q \times n} matrix of independent \eqn{q \times 1} response vectors.}

\item{A}{an \eqn{n \times m} between-individual design matrix.}

\item{M}{matrix of eigenvectors of \code{S}.}
}
\value{
A list containing the following components:
\describe{
\item{c}{the number of eigenvectors.}
\item{gcv}{the values of the gcv criteria.}
}
}
\description{
\code{gcv1.dim} estimates the number of eigenvectors by using generalized cross-validation criteria (cubic smoothing splines)
}
\details{
Here are the details of the function...
}
\keyword{internal}
