% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gcv2.dim}
\alias{gcv2.dim}
\title{Generalized cross-validation estimate of the number of eigenvectors (growth data model)}
\usage{
gcv2.dim(Y, A, M)
}
\arguments{
\item{Y}{a \eqn{q \times n} matrix of independent \eqn{q \times 1} response vectors.}

\item{A}{an \eqn{n \times m} between-individual design matrix.}

\item{M}{matrix of eigenvectors of the smoother matrix \eqn{S=(I_q+\alpha K)^{-1}}.}
}
\value{
A list containing the following components:
\describe{
\item{c}{the estimated number of eigenvectors.}
\item{gcv}{the gcv values at the points \eqn{c.grid}}
}
}
\description{
\code{gcv2.dim} estimates the number of eigenvectors by using generalized cross-validation criteria (growth data model)
}
\details{
Here are the details of the function...
}
\keyword{internal}
