% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcss.R
\name{mgcss}
\alias{mgcss}
\title{Testing of spline growth curve model}
\usage{
mgcss(Y, A = NULL, grp = NULL, model = "unif")
}
\arguments{
\item{Y}{a qxn matrix of response vectors. The ith column contains the q-variate response vector
of the ith individual. The jth row contains the responses of the individuals in the jth time point.}

\item{A}{an nxm between-individual design matrix. You should give either \code{A} or \code{grp}, not both.}

\item{grp}{an n-dimensional vector giving the m treatment groups of the observations.
The nxm between-individual design matrix \code{A} is constructed using the values of this vector:
The ith row of \code{A} is a unit vector with 1 in position \code{grp[i]} and zeros elsewhere.}

\item{model}{covariance structure. The default "unif" assumes the uniform covariance structure
\eqn{R = d^2 1_q1_q'+I_q}.}
}
\value{
A list containing the following components:
\describe{
\item{F.value}{the observed value of the F test statistic.}
\item{p.value}{approximate p-value.}
\item{p.perm}{permutation test p-value.}
\item{alpha}{the value of alpha.}
\item{c}{the value of c.}
\item{Ghat}{the spline fit.}
\item{Gtilde}{the approximated spline fit.}
}
}
\description{
\code{mgcss} is used to test the mean growth curves using cubic smoothing splines.
}
\details{
Here are the details of the function...
}
