% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.css}
\alias{plot.css}
\title{plot.css}
\usage{
\method{plot}{css}(
  x,
  xlab = "x",
  ylab = "y",
  xlim = c(min(x$x), max(x$x)),
  ylim = c(min(x$y), max(x$y)),
  main = "Approximate spline fit",
  both = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of class css.}

\item{xlab}{a title for the x axis.}

\item{ylab}{a title for the y axis.}

\item{xlim}{numeric vector of length 2, giving the x coordinate range.}

\item{ylim}{numeric vector of length 2, giving the y coordinate range.}

\item{main}{a main title.}

\item{both}{logical. If TRUE, both smoothing curve and observed points are plotted.
If FALSE, only smoothing curve is plotted.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot method for objects of class "css".
}
\details{
Here are the details of the function...
}
