% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.gcss}
\alias{plot.gcss}
\title{plot.gcss}
\usage{
\method{plot}{gcss}(x, t = 1:dim(x$Gtilde)[1], xlab = "Time points", ylab = "Gtilde", ...)
}
\arguments{
\item{x}{an object of class gcss.}

\item{t}{a \eqn{q}-dimensional vector of the time points.}

\item{xlab}{a title for the x axis.}

\item{ylab}{a title for the y axis.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot method for objects of class "gcss".
}
\details{
Here are the details of the function...
}
