% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roughness}
\alias{roughness}
\title{Roughness matrix}
\usage{
roughness(x)
}
\arguments{
\item{x}{a \eqn{q}-dimensional vector of the time points.}
}
\value{
A \eqn{q\times q} roughness matrix \eqn{K=\nabla\Delta^{-1}\nabla'}.
}
\description{
\code{roughness} is used to compute the roughness matrix.
}
\details{
Here are the details of the function...
}
\keyword{internal}
