% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summary.css}
\alias{summary.css}
\title{summary.css}
\usage{
\method{summary}{css}(object, ..., digits = 3)
}
\arguments{
\item{object}{an object of class gcss.}

\item{...}{further arguments passed to or from other methods.}

\item{digits}{minimal number of significant digits.}
}
\description{
summary.css is used to print the summary of the
testing of cubic smoothing splines and semi-parametric regression models
}
\details{
Here are the details of the function...
}
