% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summary.mgcss}
\alias{summary.mgcss}
\title{summary.mgcss}
\usage{
\method{summary}{mgcss}(object, ..., digits = 3)
}
\arguments{
\item{object}{an object of class gcss.}

\item{...}{further arguments passed to or from other methods.}

\item{digits}{minimal number of significant digits.}
}
\description{
summary.mgcss is used to print the summary of the
testing of spline growth curve model.
}
\details{
Here are the details of the function...
}
