% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gcv1.dim}
\alias{gcv1.dim}
\title{Generalized cross-validation (cubic smoothing splines)}
\usage{
gcv1.dim(y, Tn)
}
\arguments{
\item{y}{an \eqn{n}dimensional vector of responses.}

\item{Tn}{an \eqn{n \times n} matrix of eigenvectors of \eqn{S_{\alpha}=(I_n+\alpha K)^{-1}}.}
}
\value{
A list containing the following components:
\describe{
\item{c}{the number of eigenvectors.}
\item{gcv}{the values of the gcv criteria.}
}
}
\description{
\code{gcv1.dim} estimates the number of eigenvectors by using generalized cross-validation criteria (cubic smoothing splines)
}
\details{
Here are the details of the function...
}
\keyword{internal}
