% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.gcss}
\alias{plot.gcss}
\title{plot.gcss}
\usage{
\method{plot}{gcss}(
  x,
  t = 1:dim(x$Gtilde[[1]])[1],
  xlab = "Time points",
  ylab = "Gtilde",
  legend.plot = FALSE,
  inset.x = -0.2,
  ...
)
}
\arguments{
\item{x}{an object of class gcss2.}

\item{t}{a \eqn{q}-dimensional vector of the time points.}

\item{xlab}{a title for the x axis.}

\item{ylab}{a title for the y axis.}

\item{legend.plot}{logical. If TRUE, legend is added to the topright corner.}

\item{inset.x}{if legend.plot=TRUE, sets the distance of the legend from the right margin
as a fraction of the plot region.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot method for objects of class "gcss".
}
\details{
Here are the details of the function...
}
